[spirv] Remove resultType from debug instr constructors.
All debug instructions are expressed as OpExtInst in SPIR-V and they
should all have OpTypeVoid as their result type.

Taking this (OpTypeVoid) in the constructor of all debug instructions is
cumbersome and can be confusing for cases such as DebugLocalVariable or
DebugGlobalVariable or DebugFunction where we also need the variable
type or function type. For example: The caller of these constructors could
accidentally pass the variable type instead of OpTypeVoid.