WIP: Handle null decl of SpirvType
DebugTypeVisitor meets segfault when a decl of a SpirvType is null.
Since debug info is not critical for shader execution results and
we do not want compiler errors because of debug info, we temporarily
make a "Unknown" debug type for the case that DebugTypeVisitor can
not interpret debug type correctly. This must be deleted after we
emit all debug types correctly.