[spirv] handle member function even when it is not called
SpirvEmitter does not generate SpirvFunction and its information if the
function is not called. When we meet a member function of a composite,
it causes a problem because of the absense of the information. We handle
that case in LowerTypeVisitor and DebugTypeVisitor. Note that the debug
info extension spec does not allow DebugFunctionDeclaration for Members
operand of DebugTypeComposite. We must use DebugFunction.